# openapi.yaml
openapi: 3.1.0
info:
  title: Mega Bike API
  version: 0.0.1
  description: |
    API contract for the Mega Bike backend.

    Current implementation status:
    - POST /api/auth/login is implemented.
    - POST /api/auth/refresh is implemented.
    - POST /api/auth/logout is implemented.
    - GET /api/auth/me is implemented.
    - JWT access tokens are created on login.
    - Incoming bearer-token validation is implemented.
    - Local Docker PostgreSQL is exposed on localhost:5433.
servers:
  - url: http://localhost:8080
    description: Local development server
tags:
  - name: Auth
    description: Authentication endpoints
paths:
  /api/auth/login:
    post:
      tags:
        - Auth
      summary: Log in
      description: Authenticates a user with email and password and returns a signed JWT access token.
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
            examples:
              admin:
                summary: Dev admin login
                value:
                  email: admin@megabike.local
                  password: password
              manager:
                summary: Dev manager login
                value:
                  email: manager@megabike.local
                  password: password
              employee:
                summary: Dev employee login
                value:
                  email: employee@megabike.local
                  password: password
      responses:
        '200':
          description: Login successful.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
              examples:
                success:
                  value:
                    accessToken: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.example.signature
                    refreshToken: VcJKgk9jY7wFMV6zvJ4p5W7t8wz7qH0m4CO3LQ
                    tokenType: Bearer
                    expiresInSeconds: 900
                    expiresAt: '2026-07-10T10:00:00Z'
                    refreshTokenExpiresAt: '2026-07-17T10:00:00Z'
                    email: admin@megabike.local
                    authorities:
                      - PRODUCT_READ
                      - PRODUCT_WRITE
                      - ROLE_ADMIN
                      - USER_MANAGE
        '400':
          description: Request body validation failed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Invalid email or password.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                invalidCredentials:
                  value:
                    code: INVALID_CREDENTIALS
                    message: Invalid credentials.
  /api/auth/refresh:
    post:
      tags:
        - Auth
      summary: Refresh access token
      description: Issues a new access token from a valid, unexpired, non-revoked refresh token.
      operationId: refreshToken
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshTokenRequest'
      responses:
        '200':
          description: New access token issued.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '400':
          description: Request body validation failed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Invalid, expired, or revoked refresh token.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/auth/logout:
    post:
      tags:
        - Auth
      summary: Log out
      description: Revokes a refresh token. Requires a valid bearer access token.
      operationId: logout
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LogoutRequest'
      responses:
        '200':
          description: Logout processed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LogoutResponse'
        '400':
          description: Request body validation failed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Missing, invalid, or expired bearer token.
  /api/auth/me:
    get:
      tags:
        - Auth
      summary: Get current user
      description: Returns the currently authenticated user from the bearer access token.
      operationId: currentUser
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Current authenticated user.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CurrentUserResponse'
              examples:
                admin:
                  value:
                    email: admin@megabike.local
                    authorities:
                      - PRODUCT_READ
                      - PRODUCT_WRITE
                      - ROLE_ADMIN
                      - USER_MANAGE
        '403':
          description: Missing, invalid, or expired bearer token.
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    LoginRequest:
      type: object
      additionalProperties: false
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          maxLength: 320
          example: admin@megabike.local
        password:
          type: string
          format: password
          example: password
    LoginResponse:
      type: object
      additionalProperties: false
      required:
        - accessToken
        - refreshToken
        - tokenType
        - expiresInSeconds
        - expiresAt
        - refreshTokenExpiresAt
        - email
        - authorities
      properties:
        accessToken:
          type: string
          description: Signed JWT access token.
        refreshToken:
          type: string
          description: Opaque refresh token. Store securely; the database stores only its SHA-256 hash.
        tokenType:
          type: string
          enum:
            - Bearer
        expiresInSeconds:
          type: integer
          format: int64
          example: 900
        expiresAt:
          type: string
          format: date-time
        refreshTokenExpiresAt:
          type: string
          format: date-time
        email:
          type: string
          format: email
        authorities:
          type: array
          items:
            type: string
          example:
            - ROLE_ADMIN
            - PRODUCT_READ
            - PRODUCT_WRITE
    RefreshTokenRequest:
      type: object
      additionalProperties: false
      required:
        - refreshToken
      properties:
        refreshToken:
          type: string
    LogoutRequest:
      type: object
      additionalProperties: false
      required:
        - refreshToken
      properties:
        refreshToken:
          type: string
    TokenResponse:
      type: object
      additionalProperties: false
      required:
        - accessToken
        - tokenType
        - expiresInSeconds
        - expiresAt
      properties:
        accessToken:
          type: string
        tokenType:
          type: string
          enum:
            - Bearer
        expiresInSeconds:
          type: integer
          format: int64
          example: 900
        expiresAt:
          type: string
          format: date-time
    LogoutResponse:
      type: object
      additionalProperties: false
      required:
        - revoked
      properties:
        revoked:
          type: boolean
    CurrentUserResponse:
      type: object
      additionalProperties: false
      required:
        - email
        - authorities
      properties:
        email:
          type: string
          format: email
        authorities:
          type: array
          items:
            type: string
          example:
            - ROLE_ADMIN
            - PRODUCT_READ
            - PRODUCT_WRITE
    ErrorResponse:
      type: object
      additionalProperties: false
      required:
        - code
        - message
        - details
        - timestamp
      properties:
        code:
          type: string
          example: INVALID_CREDENTIALS
        message:
          type: string
          example: Invalid credentials.
        details:
          type: object
          additionalProperties:
            type: string
          example:
            email: must be a well-formed email address
        timestamp:
          type: string
          format: date-time